🔄 refactor: 이전에 코딩한 숏로그 피드 컴포넌트들 일부 common으로 이동
